\name{EMeth-package}
\alias{EMeth-package}
\alias{EMeth}
\docType{package}
\title{
\packageTitle{EMeth}
}
\description{
\packageDescription{EMeth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMeth}
\packageIndices{EMeth}
This package implements cell type deconvolution by EM algorithm based on methylation data. Including "deconvEM","deconvEM_CV" functions.
}
\author{
\packageAuthor{EMeth}

Maintainer: \packageMaintainer{EMeth}
}
\references{
TBDone
\keyword{ package }
\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ See Github pages for examples. ~~
}
