\name{EMeth-package}
\alias{EMeth-package}
\alias{EMeth}
\docType{package}
\title{
\packageTitle{EMeth}
}
\description{
\packageDescription{EMeth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMeth}
\packageIndices{EMeth}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EMeth}

Maintainer: \packageMaintainer{EMeth}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
