\name{EMeth-package}
\alias{EMeth-package}
\alias{EMeth}
\docType{package}
\title{
\packageTitle{EMeth}
}
\description{
\packageDescription{EMeth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMeth}
\packageIndices{EMeth}
~~ This package provides implementations of the EMeth algorithm. It contains two important functions.
- emeth, which provides implementation of two families of EMeth, normal and laplace, accordingto what likelihood is used. Generally we recommend the use of laplace family.
- cv.emeth, which helps tuning the ridge penalty by cross validation ~~
}
\author{
\packageAuthor{EMeth}

Maintainer: \packageMaintainer{EMeth}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
# Examples can be found in the "example" subdirectory.
}
